+++ frontend/src/App.tsx
@@ -17,6 +17,7 @@ import { HodMomoProvider } from './hod_momo/HodMomoProvider';
 import { ScannerDataProvider } from './scanner/ScannerDataContext';
 import { IbkrAccountProvider } from './ibkr/IbkrAccountContext';
 import { GatewayDisconnectedBannerHost } from './ibkr/GatewayDisconnectedBannerHost';
+import { MwcbBannerHost } from './ibkr/MwcbBannerHost';
 import { TradingPrerequisitesGate } from './ibkr/TradingPrerequisitesGate';
 import { DashboardPage } from './pages/DashboardPage';
 import { isSampleView } from './sample_data/sampleNav';
@@ -71,6 +72,7 @@ function AppShell() {
             <GlobalAppBar />
             <TradingPrerequisitesGate />
             <GatewayDisconnectedBannerHost />
+            <MwcbBannerHost />
             <NovaOsAttentionStrip global />
             <div className="nova-app-branch">
               <TraderDockLayer />